+++ src/components/anchor/anchor.stories.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Anchor, AnchorProps } from '.';
 import { Story, Meta } from '@storybook/react';
-import { Box, Flex } from '../layout';
 
 export default {
   title: 'Components/Anchor',