+++ src/components/badge/badge.stories.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Badge, BadgeProps } from '.';
 import { Story, Meta } from '@storybook/react';
-import { Box, Flex } from '../layout';
+import { Flex } from '../layout';
+import { Box } from '../box';
 
 export default {
   title: 'Components/Badge',