+++ src/components/box/box.test.tsx
@@ -1,0 +1,7 @@
+import { Box } from '.';
+
+describe('Box', () => {
+  it('is truthy', () => {
+    expect(Box).toBeTruthy();
+  });
+});