+++ src/components/index.ts
@@ -1,6 +1,7 @@
 export * from './anchor';
 export * from './avatar';
 export * from './badge';
+export * from './box';
 export * from './button';
 export * from './camara-theme';
 export * from './card';
@@ -8,7 +9,6 @@ export * from './divider';
 export * from './form';
 export * from './image';
 export * from './input';
-export * from './layout';
 export * from './layout';
 export * from './layout';
 export * from './layout';