+++ src/components/layout/index.ts
@@ -1,4 +1,3 @@
-export * from './box';
 export * from './flex';
 export * from './spacer';
 export * from './layout';